Normalise the construction and syntax of exception messages within
storageadmin and its associated unit tests so that they adhere to the
following rules:

- Exception message strings and their consumer counterpart comparisons
(eg: unit tests and smart_manager’s replication) are to be formatted
using string.format.

- Each error message starts with a capital letter and ends in a full
stop.

- Where brackets are used to for example display a pool / share name we
surround the brackets with spaces. Ie not: "Pool(leon) already exists."
but: "Pool (leon) already exists.": eases readability.

- Numbers are expressed in their numeric variant where appropriate: ie
'4' not 'Four': eases translation.

- Where a space is to exist in a multi part string we put this on the
trailing end of the prior element.

- Where a standard unit designation is used, ie 'GB', we precede it with
a space ie "6.7 GB".

- When quoting a system message at the end of our error message string
we surround that message in brackets "()" so as to discern our full stop
from any it may have.

- 'Pool', 'share', 'snapshot', 'rock-on', and 'disk' are not proper
nouns so are not to be capitalise unless used to reference Web-UI
entries ie "A Disks page Rescan may help." where both 'Disks' and
'Rescan' reference UI elements that are also Capitalised.

Begin updating existing storageadmin unit test fixtures re mocking and
exception message interpretation where necessary as they should help to
prevent the observed regression type going forward.

Unit test restoration informed additional model.save() additions in
views share.py and snapshot.py plus the addition of a conditional
delete() in nfs_export.py.

Specified CommandException (in osi.py) to prevent PyCharm’s "No
exception to reraise" error which hampers that IDE’s function re code
analysis. The resulting code was tested to ensure it’s behaviour was as
expected.

Remove build blocking gulp-eslint step.

+++ src/rockstor/storageadmin/views/nfs_exports.py
@@ -206,7 +206,9 @@ class NFSExportGroupDetailView(NFSExportMixin, rfc.GenericView):
                 nfs4_mount_teardown(export_pt)
                 cur_exports.remove(e)
                 e.delete()
-            eg.delete()
+            # Following conditional delete was informed by test_nfs_export.py:
+            if eg.id is not None:
+                eg.delete()
             exports = self.create_nfs_export_input(cur_exports)
             adv_entries = [e.export_str for e in
                            AdvancedNFSExport.objects.all()]
